requirements.txt
pillow
pytesseract
tesseract-ocr
transformers
langchain
torch